bank/requirements.txt
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
djangorestframework_simplejwt==5.5.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0 
django-prometheus==2.3.1